Clean up code.
--HG--
extra : rebase_source : db31b49e8a66a93895572be29d1c857ded702ee5

+++ Core/Utility/MultipartWebRequest.cs
@@ -74,7 +74,7 @@ namespace NuGet
             request.ContentLength = byteContent.Length;
             using (Stream requestStream = request.GetRequestStream())
             {
-                requestStream.Write(byteContent, 0, (int)byteContent.Length);
+                requestStream.Write(byteContent, 0, byteContent.Length);
             }
         }
